Fix em dash encoding and equation
The source of your problem creating the PDF is that you must have written the document in Word and then pasted it into the .Rmd file. This introduced em dashes (`-`) in an encoding that LaTeX doesn't understand. We will actually talk about character encodings on Monday. But I've fixed it.

I also put your model equation in LaTeX math format so that it is cleaner. This however is not necessary to create the PDF, but is simply stylistic.